W,A,S,D Communication working
Added inputbuffer to ATMEGA2560.
Changed Struct SD: Removed steeringmode (using extern steeringmode now).
other random stuff like version numbers ect.

+++ software/arduino/src/lib/serial.h
@@ -19,7 +19,7 @@ void serialPrintCharacterSynchronous(const char);
 void serialPrintByte(const uint8_t);
 void serialPrintByteSynchronous(const uint8_t);
 
-int8_t serialAvailable();
+uint16_t serialAvailable();
 int8_t outputBufferWalked();
 void sleepUntilEmptyOutputBuffer();
 void writeCharacterFromBuffer();
